requirements.txt
ccxt
PyYAML
SQLAlchemy
requests
beautifulsoup4
openai
google-generativeai
dashscope
chromadb
ta
pytest
pytest-mock
